+++ package.json
@@ -26,7 +26,8 @@
     "babel-preset-stage-0": "^6.5.0",
     "css-loader": "^0.23.1",
     "electron-builder": "^5.7.0",
-    "electron-prebuilt": "^1.2.5",
+    "electron-prebuilt": "^1.2.7",
+    "gulp": "^3.9.1",
     "react-css-modules": "^3.7.6",
     "react-redux": "^4.4.5",
     "sass-loader": "^4.0.0",
