+++ src/components/login/index.js
@@ -1,1 +1,1 @@
-export default from './Login.jsx';
+export default from './Login';
